2010-01-29  Zoltan Varga  <vargaz@gmail.com>
	* Int32Test.cs: Fix a copy-paste error.

svn path=/trunk/mcs/; revision=150570

+++ mcs/class/corlib/Test/System/Int32Test.cs
@@ -234,10 +234,10 @@ public class Int32Test
 		
 		Int32.Parse ("123", new DateTimeFormatInfo ());
 
-		Assert.AreEqual (734561, Int64.Parse ("734561\0"), "C#43");
-		Assert.AreEqual (734561, Int64.Parse ("734561\0\0\0    \0"), "C#44");
-		Assert.AreEqual (734561, Int64.Parse ("734561\0\0\0    "), "C#45");
-		Assert.AreEqual (734561, Int64.Parse ("734561\0\0\0"), "C#46");
+		Assert.AreEqual (734561, Int32.Parse ("734561\0"), "C#43");
+		Assert.AreEqual (734561, Int32.Parse ("734561\0\0\0    \0"), "C#44");
+		Assert.AreEqual (734561, Int32.Parse ("734561\0\0\0    "), "C#45");
+		Assert.AreEqual (734561, Int32.Parse ("734561\0\0\0"), "C#46");
 	}
 
     	[Test]
